Add Search Modes & Compare Search Mode
This update discontinues the compare website and adds search modes, a new feature allowing you to have more features when searching, and adding compare as a search mode. Popular names may be moved to a search mode soon as well.
