+++ app/page.tsx
@@ -21,15 +21,17 @@ import {
 import { useState, useEffect, Suspense } from "react";
 import { motion } from "motion/react";
 import Chart from "@/components/chart";
+import CompareChart from "@/components/comparechart";
 import Link from "next/link";
 import { useSearchParams } from "next/navigation";
 import { TopBar } from "@/components/top-bar";
 // import { Checkbox } from "@/components/ui/checkbox";
 // import { Badge } from "@/components/ui/badge";
 import Heatmap from "@/components/heatmap";
-import { Download } from "lucide-react";
+import { Download, SearchIcon } from "lucide-react";
 import Actuary from "@/components/actuary";
 import Numbers from "@/components/numbers";
+import { useTheme } from "next-themes";
 const inter = Inter({ subsets: ["latin"] });
 import {
   DropdownMenu,
@@ -55,27 +57,57 @@ function Search() {
       : searchParams.get("sex")?.toLowerCase() === "female"
       ? "F"
       : "");
+  const urlName1 = searchParams.get("name1");
+  const urlSex1 =
+    searchParams.get("sex1")?.toUpperCase() ||
+    (searchParams.get("sex1")?.toLowerCase() === "male"
+      ? "M"
+      : searchParams.get("sex1")?.toLowerCase() === "female"
+      ? "F"
+      : "");
+  const urlSearchMode = searchParams.get("mode");
   // const urlActuary = searchParams.get("actuary");
   const [name, setName] = useState(urlName || "");
   const [sex, setSex] = useState(urlSex || "");
+  const [name1, setName1] = useState(urlName1 || "");
+  const [sex1, setSex1] = useState(urlSex1 || "");
+  const [searchMode, setSearchMode] = useState(urlSearchMode || "regular");
   // const [showActuary, setShowActuary] = useState(urlActuary === "true");
   const [data, setData] = useState<NameData[]>([]);
+  const [data1, setData1] = useState<NameData[]>([]);
   const [isLoading, setIsLoading] = useState(false);
   const [hasSearched, setHasSearched] = useState(false);
   const [submittedName, setSubmittedName] = useState("");
   const [submittedSex, setSubmittedSex] = useState("");
+  const [submittedName1, setSubmittedName1] = useState("");
+  const [submittedSex1, setSubmittedSex1] = useState("");
+  const { theme, setTheme } = useTheme();
 
-  const performSearch = async (searchName: string, searchSex: string) => {
+  const performSearch = async (
+    searchName: string,
+    searchSex: string,
+    isSecond = false
+  ) => {
     setIsLoading(true);
     try {
       const response = await fetch(
         `/api/names?name=${searchName}&sex=${searchSex}`
       );
       const result = await response.json();
-      setData(result);
-      setHasSearched(true);
-      setSubmittedName(searchName);
-      setSubmittedSex(searchSex);
+
+      if (isSecond) {
+        setData1(result);
+        setSubmittedName1(searchName);
+        setSubmittedSex1(searchSex);
+      } else {
+        setData(result);
+        setSubmittedName(searchName);
+        setSubmittedSex(searchSex);
+      }
+
+      if (!hasSearched) {
+        setHasSearched(true);
+      }
     } catch (error) {
       console.error("Error fetching data:", error);
     } finally {
@@ -89,13 +121,30 @@ function Search() {
       return;
     }
 
+    if (searchMode === "compare" && (!name1 || !sex1)) {
+      return;
+    }
+
     const url = new URL(window.location.href);
     url.searchParams.set("name", name);
     url.searchParams.set("sex", sex);
+    url.searchParams.set("mode", searchMode);
+
+    if (searchMode === "compare" && name1 && sex1) {
+      url.searchParams.set("name1", name1);
+      url.searchParams.set("sex1", sex1);
+    } else {
+      url.searchParams.delete("name1");
+      url.searchParams.delete("sex1");
+    }
+
     // if (showActuary) url.searchParams.set("actuary", "true");
     window.history.pushState({}, "", url);
 
     await performSearch(name, sex);
+    if (searchMode === "compare" && name1 && sex1) {
+      await performSearch(name1, sex1, true);
+    }
   };
 
   useEffect(() => {
@@ -103,38 +152,67 @@ function Search() {
       const currentUrl = new URL(window.location.href);
       const currentName = currentUrl.searchParams.get("name");
       const currentSex = currentUrl.searchParams.get("sex");
+      const currentName1 = currentUrl.searchParams.get("name1");
+      const currentSex1 = currentUrl.searchParams.get("sex1");
+      const currentMode = currentUrl.searchParams.get("mode");
 
       if (currentName && currentSex) {
         setName(currentName);
         setSex(currentSex);
+        setSearchMode(currentMode || "regular");
         performSearch(currentName, currentSex);
+
+        if (currentMode === "compare" && currentName1 && currentSex1) {
+          setName1(currentName1);
+          setSex1(currentSex1);
+          performSearch(currentName1, currentSex1, true);
+        }
       } else {
         setHasSearched(false);
         setData([]);
+        setData1([]);
       }
     };
 
     window.addEventListener("popstate", handlePopState);
     return () => window.removeEventListener("popstate", handlePopState);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   useEffect(() => {
     if (urlName && urlSex && !hasSearched) {
       setName(urlName);
       setSex(urlSex);
+      setSearchMode(urlSearchMode || "regular");
       // setShowActuary(urlActuary === "true");
       performSearch(urlName, urlSex);
+
+      if (urlSearchMode === "compare" && urlName1 && urlSex1) {
+        setName1(urlName1);
+        setSex1(urlSex1);
+        performSearch(urlName1, urlSex1, true);
+      }
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [urlName, urlSex]);
+  }, [urlName, urlSex, urlName1, urlSex1, urlSearchMode]);
 
-  const exportToCsv = () => {
-    if (data.length === 0) return;
+  const diveDeeper = (dataSet: number = 0) => {
+    const nameToExport = dataSet === 0 ? submittedName : submittedName1;
+    const sexToExport = dataSet === 0 ? submittedSex : submittedSex1;
+    const url = "https://nomen.sh?name=" + nameToExport + "&sex=" + sexToExport;
+    window.open(url, "_blank");
+  };
+  const exportToCsv = (dataSet: number = 0) => {
+    const dataToExport = dataSet === 0 ? data : data1;
+    const nameToExport = dataSet === 0 ? submittedName : submittedName1;
+    const sexToExport = dataSet === 0 ? submittedSex : submittedSex1;
+
+    if (dataToExport.length === 0) return;
 
     const headers = ["Name", "Sex", "Amount", "Year"];
     const csvContent = [
       headers.join(","),
-      ...data.map(
+      ...dataToExport.map(
         (item) => `${item.name},${item.sex},${item.amount},${item.year}`
       ),
     ].join("\n");
@@ -143,7 +221,7 @@ function Search() {
     const url = URL.createObjectURL(blob);
     const link = document.createElement("a");
     link.setAttribute("href", url);
-    link.setAttribute("download", `${submittedName}-${submittedSex}.csv`);
+    link.setAttribute("download", `${nameToExport}-${sexToExport}.csv`);
     link.style.visibility = "hidden";
     document.body.appendChild(link);
     link.click();
@@ -151,17 +229,21 @@ function Search() {
     URL.revokeObjectURL(url);
   };
 
-  const exportToJson = () => {
-    if (data.length === 0) return;
+  const exportToJson = (dataSet: number = 0) => {
+    const dataToExport = dataSet === 0 ? data : data1;
+    const nameToExport = dataSet === 0 ? submittedName : submittedName1;
+    const sexToExport = dataSet === 0 ? submittedSex : submittedSex1;
 
-    const jsonContent = JSON.stringify(data, null, 2);
+    if (dataToExport.length === 0) return;
+
+    const jsonContent = JSON.stringify(dataToExport, null, 2);
     const blob = new Blob([jsonContent], {
       type: "application/json;charset=utf-8;",
     });
     const url = URL.createObjectURL(blob);
     const link = document.createElement("a");
     link.setAttribute("href", url);
-    link.setAttribute("download", `${submittedName}-${submittedSex}.json`);
+    link.setAttribute("download", `${nameToExport}-${sexToExport}.json`);
     link.style.visibility = "hidden";
     document.body.appendChild(link);
     link.click();
@@ -169,12 +251,16 @@ function Search() {
     URL.revokeObjectURL(url);
   };
 
-  const exportToXml = () => {
-    if (data.length === 0) return;
+  const exportToXml = (dataSet: number = 0) => {
+    const dataToExport = dataSet === 0 ? data : data1;
+    const nameToExport = dataSet === 0 ? submittedName : submittedName1;
+    const sexToExport = dataSet === 0 ? submittedSex : submittedSex1;
+
+    if (dataToExport.length === 0) return;
 
     const xmlContent = `<?xml version="1.0" encoding="UTF-8"?>
 <nameData>
-${data
+${dataToExport
   .map(
     (item) => `  <record>
     <name>${item.name}</name>
@@ -192,7 +278,7 @@ ${data
     const url = URL.createObjectURL(blob);
     const link = document.createElement("a");
     link.setAttribute("href", url);
-    link.setAttribute("download", `${submittedName}-${submittedSex}.xml`);
+    link.setAttribute("download", `${nameToExport}-${sexToExport}.xml`);
     link.style.visibility = "hidden";
     document.body.appendChild(link);
     link.click();
@@ -200,9 +286,11 @@ ${data
     URL.revokeObjectURL(url);
   };
 
-  const exportToPdf = () => {
-    if (data.length === 0) return;
-  const htmlContent = `
+  const exportToPdf = (dataSet: number = 0) => {
+    const dataToExport = dataSet === 0 ? data : data1;
+
+    if (dataToExport.length === 0) return;
+    const htmlContent = `
 <!DOCTYPE html>
 <html>
 <head>
@@ -289,7 +377,7 @@ ${data
         </tr>
       </thead>
       <tbody>
-        ${data
+        ${dataToExport
           .map(
             (item) => `
           <tr>
@@ -330,40 +418,137 @@ ${data
         <div className="flex flex-col items-center justify-center min-h-screen p-6">
           <div className="w-[30rem] max-w-full">
             <motion.p className="font-bold text-2xl font-gosha lowercase pb-1">
-              Nomen
+              Nomen{" "}
+              {searchMode === "compare" && (
+                <span className="text-lg">COMPARE</span>
+              )}
             </motion.p>
             <p className="text-muted-foreground text-sm pb-5">
-              A parser for every name listed on a social security card between
-              1880-2024, tabulated from the United States Social Security
-              Adminstration{"'"}s data.
+              {searchMode === "compare"
+                ? "Compare trends between two names from 1880-2023, using data from the United States Social Security Administration."
+                : "A parser for every name listed on a social security card between 1880-2024, tabulated from the United States Social Security Adminstration's data."}
             </p>
             <form onSubmit={handleSearch} className="flex flex-col space-y-5">
-              <motion.div
-                className="flex flex-col space-y-2"
-                layoutId="name-input-container"
-              >
-                <p>First Name:</p>
-                <motion.div layoutId="name-input">
-                  <Input
-                    type="text"
-                    placeholder="Enter a name"
-                    className="w-full"
-                    aria-label="Name search"
-                    value={name}
-                    onChange={(e) => setName(e.target.value)}
-                  />
-                </motion.div>
-              </motion.div>
-              <motion.div className="" layoutId="sex-select-container">
-                <p>Sex:</p>
-                <motion.div layoutId="sex-select">
-                  <Select value={sex} onValueChange={setSex}>
+              {searchMode === "regular" ? (
+                <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+                  <motion.div
+                    className="flex flex-col space-y-2"
+                    layoutId="name-input-container"
+                  >
+                    <p>First Name:</p>
+                    <motion.div layoutId="name-input">
+                      <Input
+                        type="text"
+                        placeholder="Enter a name"
+                        className="w-full"
+                        aria-label="Name search"
+                        value={name}
+                        onChange={(e) => setName(e.target.value)}
+                      />
+                    </motion.div>
+                  </motion.div>
+                  <motion.div
+                    className="flex flex-col space-y-2"
+                    layoutId="sex-select-container"
+                  >
+                    <p>Sex:</p>
+                    <motion.div layoutId="sex-select">
+                      <Select value={sex} onValueChange={setSex}>
+                        <SelectTrigger>
+                          <SelectValue placeholder="Select gender" />
+                        </SelectTrigger>
+                        <SelectContent>
+                          <SelectItem value="M">Male</SelectItem>
+                          <SelectItem value="F">Female</SelectItem>
+                        </SelectContent>
+                      </Select>
+                    </motion.div>
+                  </motion.div>
+                </div>
+              ) : (
+                <div className="grid grid-cols-1 md:grid-cols-2 gap-5">
+                  <div className="space-y-5">
+                    <motion.div
+                      className="flex flex-col space-y-2"
+                      layoutId="name-input-container"
+                    >
+                      <p>First Name:</p>
+                      <motion.div layoutId="name-input">
+                        <Input
+                          type="text"
+                          placeholder="Enter a name"
+                          className="w-full"
+                          aria-label="Name search"
+                          value={name}
+                          onChange={(e) => setName(e.target.value)}
+                        />
+                      </motion.div>
+                    </motion.div>
+                    <motion.div
+                      className="flex flex-col space-y-2"
+                      layoutId="sex-select-container"
+                    >
+                      <p>Sex:</p>
+                      <motion.div layoutId="sex-select">
+                        <Select value={sex} onValueChange={setSex}>
+                          <SelectTrigger>
+                            <SelectValue placeholder="Select gender" />
+                          </SelectTrigger>
+                          <SelectContent>
+                            <SelectItem value="M">Male</SelectItem>
+                            <SelectItem value="F">Female</SelectItem>
+                          </SelectContent>
+                        </Select>
+                      </motion.div>
+                    </motion.div>
+                  </div>
+                  <div className="space-y-5">
+                    <motion.div
+                      className="flex flex-col space-y-2"
+                      layoutId="name1-input-container"
+                    >
+                      <p>Second Name</p>
+                      <motion.div layoutId="name1-input">
+                        <Input
+                          type="text"
+                          placeholder="Enter a name to compare"
+                          className="w-full"
+                          aria-label="Second name search"
+                          value={name1}
+                          onChange={(e) => setName1(e.target.value)}
+                        />
+                      </motion.div>
+                    </motion.div>
+                    <motion.div
+                      className="flex flex-col space-y-2"
+                      layoutId="sex1-select-container"
+                    >
+                      <p>Sex:</p>
+                      <motion.div layoutId="sex1-select">
+                        <Select value={sex1} onValueChange={setSex1}>
+                          <SelectTrigger>
+                            <SelectValue placeholder="Select gender" />
+                          </SelectTrigger>
+                          <SelectContent>
+                            <SelectItem value="M">Male</SelectItem>
+                            <SelectItem value="F">Female</SelectItem>
+                          </SelectContent>
+                        </Select>
+                      </motion.div>
+                    </motion.div>
+                  </div>
+                </div>
+              )}
+              <motion.div className="" layoutId="search-mode-select-container">
+                <p>Search Mode:</p>
+                <motion.div layoutId="search-mode-select">
+                  <Select value={searchMode} onValueChange={setSearchMode}>
                     <SelectTrigger>
-                      <SelectValue placeholder="Select gender" />
+                      <SelectValue placeholder="Select search mode" />
                     </SelectTrigger>
                     <SelectContent>
-                      <SelectItem value="M">Male</SelectItem>
-                      <SelectItem value="F">Female</SelectItem>
+                      <SelectItem value="regular">Regular</SelectItem>
+                      <SelectItem value="compare">Compare</SelectItem>
                     </SelectContent>
                   </Select>
                 </motion.div>
@@ -377,7 +562,11 @@ ${data
                     exit={{ opacity: 0, filter: "blur(4px)" }}
                     transition={{ duration: 0.6 }}
                   >
-                    {isLoading ? "Searching..." : "Search"}
+                    {isLoading
+                      ? "Searching..."
+                      : searchMode === "compare"
+                      ? "Compare"
+                      : "Search"}
                   </motion.span>
                 </Button>
               </motion.div>
@@ -386,13 +575,24 @@ ${data
                   Popular Names
                 </Link>
                 {" | "}
-                <Link href="/compare" className="underline">
-                  Compare Names
-                </Link>
-                {" | "}
                 <Link href="/about" className="underline">
                   About
                 </Link>
+                {" | "}
+                <button
+                  onClick={() =>
+                    setTheme(
+                      theme === "dark"
+                        ? "light"
+                        : theme === "system"
+                        ? "light"
+                        : "dark"
+                    )
+                  }
+                  className="underline"
+                >
+                  Change Theme
+                </button>
               </div>
             </form>
           </div>
@@ -404,38 +604,97 @@ ${data
   return (
     <>
       <div className={inter.className + " min-h-screen flex flex-col"}>
-        <TopBar>
+        <TopBar title={searchMode === "compare" ? "COMPARE" : ""}>
           <form
             onSubmit={handleSearch}
             className="flex-1 flex flex-col md:flex-row gap-4 items-center"
           >
-            <motion.div className="w-full" layoutId="name-input-container">
-              <motion.div layoutId="name-input">
-                <Input
-                  type="text"
-                  placeholder="Enter a name"
-                  aria-label="Name search"
-                  value={name}
-                  onChange={(e) => setName(e.target.value)}
-                />
-              </motion.div>
-            </motion.div>
-            <motion.div
-              className="w-full md:w-40"
-              layoutId="sex-select-container"
-            >
-              <motion.div layoutId="sex-select">
-                <Select value={sex} onValueChange={setSex}>
-                  <SelectTrigger>
-                    <SelectValue placeholder="Select gender" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="M">Male</SelectItem>
-                    <SelectItem value="F">Female</SelectItem>
-                  </SelectContent>
-                </Select>
-              </motion.div>
-            </motion.div>
+            {searchMode === "regular" ? (
+              <>
+                <motion.div className="w-full" layoutId="name-input-container">
+                  <motion.div layoutId="name-input">
+                    <Input
+                      type="text"
+                      placeholder="Enter a name"
+                      aria-label="Name search"
+                      value={name}
+                      onChange={(e) => setName(e.target.value)}
+                    />
+                  </motion.div>
+                </motion.div>
+                <motion.div
+                  className="w-full md:w-40"
+                  layoutId="sex-select-container"
+                >
+                  <motion.div layoutId="sex-select">
+                    <Select value={sex} onValueChange={setSex}>
+                      <SelectTrigger>
+                        <SelectValue placeholder="Select gender" />
+                      </SelectTrigger>
+                      <SelectContent>
+                        <SelectItem value="M">Male</SelectItem>
+                        <SelectItem value="F">Female</SelectItem>
+                      </SelectContent>
+                    </Select>
+                  </motion.div>
+                </motion.div>
+              </>
+            ) : (
+              <>
+                <div className="w-full md:w-1/3 grid grid-cols-2 gap-2">
+                  <motion.div layoutId="name-input-container">
+                    <motion.div layoutId="name-input">
+                      <Input
+                        type="text"
+                        placeholder="First name"
+                        aria-label="Name search"
+                        value={name}
+                        onChange={(e) => setName(e.target.value)}
+                      />
+                    </motion.div>
+                  </motion.div>
+                  <motion.div layoutId="sex-select-container">
+                    <motion.div layoutId="sex-select">
+                      <Select value={sex} onValueChange={setSex}>
+                        <SelectTrigger>
+                          <SelectValue placeholder="Sex" />
+                        </SelectTrigger>
+                        <SelectContent>
+                          <SelectItem value="M">Male</SelectItem>
+                          <SelectItem value="F">Female</SelectItem>
+                        </SelectContent>
+                      </Select>
+                    </motion.div>
+                  </motion.div>
+                </div>
+                <div className="w-full md:w-1/3 grid grid-cols-2 gap-2">
+                  <motion.div layoutId="name1-input-container">
+                    <motion.div layoutId="name1-input">
+                      <Input
+                        type="text"
+                        placeholder="Second name"
+                        aria-label="Second name search"
+                        value={name1}
+                        onChange={(e) => setName1(e.target.value)}
+                      />
+                    </motion.div>
+                  </motion.div>
+                  <motion.div layoutId="sex1-select-container">
+                    <motion.div layoutId="sex1-select">
+                      <Select value={sex1} onValueChange={setSex1}>
+                        <SelectTrigger>
+                          <SelectValue placeholder="Sex" />
+                        </SelectTrigger>
+                        <SelectContent>
+                          <SelectItem value="M">Male</SelectItem>
+                          <SelectItem value="F">Female</SelectItem>
+                        </SelectContent>
+                      </Select>
+                    </motion.div>
+                  </motion.div>
+                </div>
+              </>
+            )}
             <motion.div layoutId="search-button">
               <Button type="submit" disabled={isLoading}>
                 <motion.span
@@ -445,33 +704,237 @@ ${data
                   exit={{ opacity: 0, filter: "blur(4px)" }}
                   transition={{ duration: 0.6 }}
                 >
-                  {isLoading ? "Searching..." : "Search"}
+                  {isLoading
+                    ? "Searching..."
+                    : searchMode === "compare"
+                    ? "Compare"
+                    : "Search"}
                 </motion.span>
               </Button>
             </motion.div>
           </form>
         </TopBar>
         {data.length > 0 && (
-          <div className="pt-3 px-2 sm:pt-5 sm:px-9 grid grid-cols-1 md:grid-cols-2 gap-3 md:gap-2">
-            <Chart name={submittedName} sex={submittedSex} />
-            <Heatmap sex={submittedSex} name={submittedName} />
-            <>
-              <div>
-                <Actuary name={submittedName} sex={submittedSex} />
-              </div>
-              <div>
-                <Numbers name={submittedName} sex={submittedSex} />
-              </div>
-            </>
+          <div className="pt-3 px-2 sm:pt-5 sm:px-9 grid grid-cols-1 gap-3 md:gap-2">
+            {searchMode === "compare" ? (
+              <CompareChart
+                name={submittedName}
+                sex={submittedSex}
+                name1={submittedName1}
+                sex1={submittedSex1}
+              />
+            ) : (
+              <>
+                <div className="grid grid-cols-1 md:grid-cols-2 gap-3 md:gap-2">
+                  <Chart name={submittedName} sex={submittedSex} />
+                  <Heatmap sex={submittedSex} name={submittedName} />
+                </div>
+                <div className="grid grid-cols-1 md:grid-cols-2 gap-3 md:gap-2">
+                  <Actuary name={submittedName} sex={submittedSex} />
+                  <Numbers name={submittedName} sex={submittedSex} />
+                </div>
+              </>
+            )}
           </div>
         )}
-        <div className="flex-1 overflow-auto p-4">
-          {data.length > 0 ? (
+        {searchMode === "compare" && (
+          <div className="md:hidden p-4 text-center">
+            <div className="bg-secondary/30 rounded-lg p-4">
+              <p className="text-muted-foreground">
+                Full data access is not compatible on mobile at the moment
+              </p>
+              <p className="text-xs mt-2">
+                Please view on a larger screen for detailed tables
+              </p>
+            </div>
+          </div>
+        )}
+
+        <div
+          className={
+            searchMode === "compare"
+              ? "hidden md:flex flex-row space-x-9 p-9"
+              : "flex-1 overflow-auto p-4"
+          }
+        >
+          {searchMode === "compare" ? (
+            <>
+              <div className="flex-1 overflow-auto">
+                {data.length > 0 ? (
+                  <div className="container mx-auto">
+                    <div className="flex justify-between items-center mb-3">
+                      <h3 className="text-md font-medium pl-2">
+                        {submittedName} ({submittedSex})
+                      </h3>
+                      <div className="flex items-center gap-2">
+                        <Button
+                          variant="outline"
+                          size="sm"
+                          className="flex items-center gap-2"
+                          onClick={() => diveDeeper(0)}
+                        >
+                          <SearchIcon size={16} />
+                          Dive Deeper
+                        </Button>
+                        <DropdownMenu>
+                          <DropdownMenuTrigger>
+                            <Button
+                              variant="outline"
+                              size="sm"
+                              className="flex items-center gap-2"
+                            >
+                              <Download size={16} />
+                              Export
+                            </Button>
+                          </DropdownMenuTrigger>
+                          <DropdownMenuContent>
+                            <DropdownMenuItem onClick={() => exportToCsv(0)}>
+                              Download CSV
+                            </DropdownMenuItem>
+                            <DropdownMenuItem onClick={() => exportToPdf(0)}>
+                              Download PDF
+                            </DropdownMenuItem>
+                            <DropdownMenuItem onClick={() => exportToJson(0)}>
+                              Download JSON
+                            </DropdownMenuItem>
+                            <DropdownMenuItem onClick={() => exportToXml(0)}>
+                              Download XML
+                            </DropdownMenuItem>
+                          </DropdownMenuContent>
+                        </DropdownMenu>
+                      </div>
+                    </div>
+                    <Table>
+                      <TableHeader className="sticky top-0 bg-background">
+                        <TableRow>
+                          <TableHead className="w-[100px]">Name</TableHead>
+                          <TableHead>Sex</TableHead>
+                          <TableHead>Amount</TableHead>
+                          <TableHead className="text-right">Year</TableHead>
+                        </TableRow>
+                      </TableHeader>
+                      <TableBody>
+                        {data.map((item, index) => (
+                          <TableRow key={index}>
+                            <TableCell className="font-medium">
+                              {item.name}
+                            </TableCell>
+                            <TableCell>{item.sex}</TableCell>
+                            <TableCell>{item.amount}</TableCell>
+                            <TableCell className="text-right">
+                              {item.year}
+                            </TableCell>
+                          </TableRow>
+                        ))}
+                      </TableBody>
+                    </Table>
+                  </div>
+                ) : (
+                  <div className="h-full flex items-center justify-center flex-col">
+                    <p className="text-muted-foreground">No results found</p>
+                    <p className="text-gray-400 text-sm">
+                      Years with under five occurances are not shown for privacy
+                      reasons
+                    </p>
+                  </div>
+                )}
+              </div>
+
+              <div className="flex-1 overflow-auto">
+                {data1.length > 0 ? (
+                  <div className="container mx-auto">
+                    <div className="flex justify-between items-center mb-3">
+                      <h3 className="text-md font-medium pl-2">
+                        {submittedName1} ({submittedSex1})
+                      </h3>
+                      <div className="flex items-center gap-2">
+                        <Button
+                          variant="outline"
+                          size="sm"
+                          className="flex items-center gap-2"
+                          onClick={() => diveDeeper(1)}
+                        >
+                          <SearchIcon size={16} />
+                          Dive Deeper
+                        </Button>
+                        <DropdownMenu>
+                          <DropdownMenuTrigger>
+                            <Button
+                              variant="outline"
+                              size="sm"
+                              className="flex items-center gap-2"
+                            >
+                              <Download size={16} />
+                              Export
+                            </Button>
+                          </DropdownMenuTrigger>
+                          <DropdownMenuContent>
+                            <DropdownMenuItem onClick={() => exportToCsv(1)}>
+                              Download CSV
+                            </DropdownMenuItem>
+                            <DropdownMenuItem onClick={() => exportToPdf(1)}>
+                              Download PDF
+                            </DropdownMenuItem>
+                            <DropdownMenuItem onClick={() => exportToJson(1)}>
+                              Download JSON
+                            </DropdownMenuItem>
+                            <DropdownMenuItem onClick={() => exportToXml(1)}>
+                              Download XML
+                            </DropdownMenuItem>
+                          </DropdownMenuContent>
+                        </DropdownMenu>
+                      </div>
+                    </div>
+                    <Table>
+                      <TableHeader className="sticky top-0 bg-background">
+                        <TableRow>
+                          <TableHead className="w-[100px]">Name</TableHead>
+                          <TableHead>Sex</TableHead>
+                          <TableHead>Amount</TableHead>
+                          <TableHead className="text-right">Year</TableHead>
+                        </TableRow>
+                      </TableHeader>
+                      <TableBody>
+                        {data1.map((item, index) => (
+                          <TableRow key={index}>
+                            <TableCell className="font-medium">
+                              {item.name}
+                            </TableCell>
+                            <TableCell>{item.sex}</TableCell>
+                            <TableCell>{item.amount}</TableCell>
+                            <TableCell className="text-right">
+                              {item.year}
+                            </TableCell>
+                          </TableRow>
+                        ))}
+                      </TableBody>
+                    </Table>
+                  </div>
+                ) : submittedName1 ? (
+                  <div className="h-full flex items-center justify-center flex-col">
+                    <p className="text-muted-foreground">No results found</p>
+                    <p className="text-gray-400 text-sm">
+                      Years with under five occurances are not shown for privacy
+                      reasons
+                    </p>
+                  </div>
+                ) : (
+                  <div className="h-full flex items-center justify-center flex-col">
+                    <p className="text-muted-foreground">
+                      No second name selected
+                    </p>
+                    <p className="text-gray-400 text-sm">
+                      Add a second name to compare data
+                    </p>
+                  </div>
+                )}
+              </div>
+            </>
+          ) : data.length > 0 ? (
             <div className="container mx-auto">
               <div className="flex justify-end mb-3">
                 <DropdownMenu>
                   <DropdownMenuTrigger>
-                    {" "}
                     <Button
                       variant="outline"
                       size="sm"
@@ -482,16 +945,16 @@ ${data
                     </Button>
                   </DropdownMenuTrigger>
                   <DropdownMenuContent>
-                    <DropdownMenuItem onClick={exportToCsv}>
+                    <DropdownMenuItem onClick={() => exportToCsv(0)}>
                       Download CSV
                     </DropdownMenuItem>
-                    <DropdownMenuItem onClick={exportToPdf}>
+                    <DropdownMenuItem onClick={() => exportToPdf(0)}>
                       Download PDF
                     </DropdownMenuItem>
-                    <DropdownMenuItem onClick={exportToJson}>
+                    <DropdownMenuItem onClick={() => exportToJson(0)}>
                       Download JSON
                     </DropdownMenuItem>
-                    <DropdownMenuItem onClick={exportToXml}>
+                    <DropdownMenuItem onClick={() => exportToXml(0)}>
                       Download XML
                     </DropdownMenuItem>
                   </DropdownMenuContent>
